refactor(config): ScannerCommands 按 full build tag 拆分
