+++ core/config/scanner.go
@@ -14,17 +14,6 @@ var ExtraSummaryEntries []string
 
 var ExtraScannerUsage = map[string]func() string{}
 
-type ScannerCommands struct {
-	Scan     struct{} `command:"scan" description:"Run the scan pipeline"`
-	Gogo     struct{} `command:"gogo" description:"Run gogo scanner"`
-	Spray    struct{} `command:"spray" description:"Run spray scanner"`
-	Katana   struct{} `command:"katana" description:"Run katana web crawler"`
-	Zombie   struct{} `command:"zombie" description:"Run zombie weakpass scanner"`
-	Neutron  struct{} `command:"neutron" description:"Run neutron POC scanner"`
-	Cyberhub struct{} `command:"cyberhub" description:"Search Cyberhub fingerprints and POCs"`
-	Passive  struct{} `command:"passive" description:"Run passive cyberspace recon"`
-}
-
 func ScannerCommandAvailable(name string) bool {
 	switch name {
 	case "scan", "gogo", "spray", "zombie", "neutron", "cyberhub":
